refactor(Interaction): Remove abstract method interact()
This method can be replaced by the constructor. So it has now been
removed. Some of the extend error checked by IDE will also be solved.

BREAKING CHANGE: This commit removes a abstract method in a abstract
class so the class which extends this class will no more required to
write this method